+++ langdb/__init__.py
@@ -1,0 +1,8 @@
+"""Utility package for the language learning content database."""
+
+__all__ = [
+    "ingest",
+    "manage_words",
+    "recommend",
+    "recommend_known",
+]
